Do not store the result of init in context
C fuse allows for the result of init to be stored
as the user_data that is passed in the context
to each file operation and was also passed back as
user_data to destroy method. We now use this to store a ref
to the fuse object itself.
NOTE incompatible change
  was: destroy(context,user_data)
  now: destroy(context)
If you need to keep user data on the fuse object you can
do it directly in your filesystem class.